Viewer: Split of App component for easier testing
Divided App component to ResultViewer component to be able to easier
test result viewer.
